+++ actionmailer/test/log_subscriber_test.rb
@@ -50,16 +50,4 @@ class AMLogSubscriberTest < ActionMailer::TestCase
   ensure
     BaseMailer.deliveries.clear
   end
-
-  def test_receive_is_notified
-    fixture = File.read(File.expand_path("fixtures/raw_email", __dir__))
-    assert_deprecated do
-      TestMailer.receive(fixture)
-    end
-    wait
-    assert_equal(1, @logger.logged(:info).size)
-    assert_match(/Received mail/, @logger.logged(:info).first)
-    assert_equal(1, @logger.logged(:debug).size)
-    assert_match(/Jamis/, @logger.logged(:debug).first)
-  end
 end