+++ actionmailer/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Remove deprecated `ActionMailer::Base.receive` in favor of [Action Mailbox](https://github.com/rails/rails/tree/master/actionmailbox).
+
+    *Rafael Mendonça França*
+
 *   Fix ActionMailer assertions don't work for parameterized mail with legacy delivery job.
 
     *bogdanvlviv*